improve code quality with some handle error edit
