+++ src/main/java/Dukes/DukeException.java
@@ -1,5 +1,7 @@
-/*public class DukeException extends Exception {
+package Dukes;
+
+public class DukeException extends Exception {
         public DukeException(String message) {
             super(message);
         }
-    }*/
+    }
