+++ text-ui-test/EXPECTED.TXT
@@ -5,7 +5,7 @@ Hello from
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 
-Hello! I'm Duke
+Hello! I'm Dukes.Duke
 What can I do for you?
 
 Got it, I've added the following task:
